refactor: harden session step surface contract
